fix: prevent trips pending match queue from getting stuck (#31)
The 'Trip matching' mechanism relies on a queue to determine the next trip to match. Currently, the logic only tries to match the top/head of the queue and if match can't be done (Example: Vehicle doesn't support trip type), then the queue gets stuck. This fixes the problem by moving through the queue looking for matches.

Removes a few unused variables/imports from different files